Adjust the wwan, watchdog and alsa-loopback test cases. (BugFix) (#821)
* The commit is for the alsa-loopback test case delete.

Actually, the alsa-loopback-automated test case does the same
thing with alsa-loopback. When we test iot project machines,
the alsa-loopback-automated needs to be tested. I guess we
could do the alsa-loopback-automated only in the futuer.

* This commit to modify the watchdog test cases.

The watchdog-automated's test items have included
watchdog-maunal,  so I remove the watchdog-manual
test case.

* Modify the wwan related test cases.

The mobilebroadband-cert-automated test case is only
part of the wwan test plan. We hope to replace
mobilebroadband-cert-automated with wwan-automated.

The relevant test plans we currently use include:
client-cert-desktop-22-04-automated
client-cert-desktop-20-04-automated
client-cert-desktop-18-04-automated

I have replaced mobilebroadband-cert-automated in them
with wwan-automated, and added after-sudpend-wwan-
automated to ensure that wwan is tested before and after
suspend.

In addition, for mobilebroadband-cert-manual in the Manual
test plan, when the SIM is inserted into the prototype in the
desktop, the system should automatically recognize wwan and
SIM card, and we do not need to perform manual testing. So I
deleted mobilebroadband-cert-manual in the relevant test
plans. At the same time, since wwan-automated already
contains wwan-manual cases, I deleted the wwan-manual in
these test plans:
client-cert-desktop-22-04-manual
client-cert-desktop-20-04-manual
client-cert-desktop-18-04-manual

* Revert wwan test-plan.pxu. Keep the necessary dependence
of test cases.