fprintd should enroll a user, not a uid (BugFix) (#813)
fprintd should enroll to a user while not uid. (BugFix) (#813)
$ sudo env | grep SUDO
SUDO_COMMAND=/usr/bin/env
SUDO_USER=ubuntu
SUDO_UID=1001
SUDO_GID=1001

when running checkbox with sudo, it will 'fprintd-enroll 1001'
and the test case will lock screen to let 'ubuntu' user to login,
but at that time, it will not work because the user is not 'ubuntu'.