Implement explain() in TemplateUnit
This method overrides UnitWithId.explain() which displays the Unit's id.

In the case of a Template, this is misleading, and the template_id (or,
in this case, template_partial_id to prevent displaying the provider
namespace) should be used instead.

This is useful when reporting errors using `./manage.py validate`, for
instance.