Improve suspend/suspend_advanced* tests #857
rtwake can sometimes fail to set the alarm. While running the command
in background before suspending the system, in such a case the whole
test plan will hang.

This patch change the mode to allow the command to return immediately,
while still setting the alarm, which allow to serialize its result.