Add template-summary and template-description fields and validators
Add two fields for the Template Unit:

- template-summary
- template-description

These fields mimick the summary and description fields of the Job Unit
to provide a quick explanation and a longer form one about what a
template unit does.

Both these fields are translatable, and validation is added for the
template-summary field to make sure if a value is provided, it is one
line only and is less than 80 characters.