Refactor qualifiers.select_jobs() function
The original implementation operates on integers representing indices of
a particular job in the list of jobs. It makes the code hard to read for
a gain that is not very interesting, since this function is usually
called 2 or 3 times during a regular Checkbox run.

The following refactor does two things:

1. Operate on the jobs themsleves instead of the integers of the indices
pointing to them
2. Remove code duplication by using a nested definition, _handle_vote(),
to make the code easier to read