Search for matches in template_id in addition to the job's id field
Test plans can now include template ids in addition to job ids.

Because it's impossible to know if an item in the test plan is a job or
a template, Checkbox populates the list of qualifiers with
FieldQualifier instances looking for a match for the `id` field.

When select_jobs(job_list, qualifier_list) is called, it will first
check if the job id matches the value of the qualifier, as before. If it
does not, it will check if it matches the origin of this job (its
template_id). If this matches, then it means the FieldQualifier being
used should be modified to search for a match in the `template_id` field
instead of the `id` field.

In order to achieve this, the FieldQualifier.field needs a setter.