Allow template ids to be used in Test Plans (New) (#1009)
* Refactor qualifiers.select_jobs() function

The original implementation operates on integers representing indices of
a particular job in the list of jobs. It makes the code hard to read for
a gain that is not very interesting, since this function is usually
called 2 or 3 times during a regular Checkbox run.

The following refactor does two things:

1. Operate on the jobs themsleves instead of the integers of the indices
pointing to them
2. Remove code duplication by using a nested definition, _handle_vote(),
to make the code easier to read

* Search for matches in template_id in addition to the job's id field

Test plans can now include template ids in addition to job ids.

Because it's impossible to know if an item in the test plan is a job or
a template, Checkbox populates the list of qualifiers with
FieldQualifier instances looking for a match for the `id` field.

When select_jobs(job_list, qualifier_list) is called, it will first
check if the job id matches the value of the qualifier, as before. If it
does not, it will check if it matches the origin of this job (its
template_id). If this matches, then it means the FieldQualifier being
used should be modified to search for a match in the `template_id` field
instead of the `id` field.

In order to achieve this, the FieldQualifier.field needs a setter.

* Add unit tests for select_jobs() when template ids are used

* Update Test Plan Unit reference page with the use of template ids

* Cleanup Test Plan Unit reference page

* Add more unit tests for select_jobs()

* Remove useless condition branch

* Add comment to explain why a list of included jobs is required