Ensure suspend job only depends on jobs in the run list
If the filtering is done on the whole job_list, the suspend job will
depend on jobs that are not required (undesired) for the current session
run.

As a result, the suspend job is skipped with reasons like:

    required dependency 'com.canonical.certification::audio/
    playback_auto' did not run yet

and all of the after-suspend jobs are skipped!

the job_list therefore needs to be narrowed down to remove the
"undesired" jobs and only keep the ones that are actually going to run.