Introduce "list-testplan" subcommand
The new subcommand works similarly to "list-bootstrapped" since it also
bootstraps a given test plan.

However, it filters out the instantiated jobs and replaces them with
their original template.

In the end, a list of jobs and templates used in this test plan are
returned as JSON-formatted data.

This is to be used by external tools, for instance to create a document
listing information (summary, description...) about executed jobs and
templates.

Fix CHECKBOX-1263