Add 'expand' subcommand to list jobs and templates in a test plan
The new subcommand "expands" a given test plan by going through all its
sections and returning the pure (non-instantiated) jobs and templates
being used.

It is different from the "list-boostrapped" subcommand since it does not
perform any bootstrap.

By default, it outputs the jobs/templates id as text, but it can be
called with the `--format json` option to output JSON data with all the
known information for said jobs and templates.

This is to be used by external tools, for instance to create a document
listing information (summary, description...) about executed jobs and
templates.

Fix CHECKBOX-1263