Provide more information in Snapd._poll_change()
In order to get more data in the job logs, Snapd._poll_change() is
amended to do the following:

- Outuput info about the task when it's in the "Wait" status (and
return, as it likely means the snap is waiting for a reboot)
- Outuput info about the task when it's in the "Error" status
- If an ongoing task (status = "Doing") has a progress label assigned,
print the current progress in % (similar to what's done in the snap
command itself). This is useful to investigate if tasks like snap
downloads have frozen at some point.

Unit tests for the _poll_change() method are also added.