Make sure the device reboots after snap refresh/revert tests (BugFix) (#1124)
* Provide more information in Snapd._poll_change()

In order to get more data in the job logs, Snapd._poll_change() is
amended to do the following:

- Outuput info about the task when it's in the "Wait" status (and
return, as it likely means the snap is waiting for a reboot)
- Outuput info about the task when it's in the "Error" status
- If an ongoing task (status = "Doing") has a progress label assigned,
print the current progress in % (similar to what's done in the snap
command itself). This is useful to investigate if tasks like snap
downloads have frozen at some point.

Unit tests for the _poll_change() method are also added.

* Return the response when calling Snapd.install() and remove()

In order to be more consistent with other methods, these methods now
return the response.

Unit tests are added for these methods as well.

* Add timeout argument to snap_update_test.py and output more info

Script can now be called with the --timeout argument (defaulting to 300
seconds).

This is passed to the Snapd() instance, along with the verbosity flag
that enables Snapd._info() to output useful information for the job
logs.

* Replace logging with print statements in snap_update_test.py

Print statements make more sense because any information captured during
the snap commands should be made available to the Checkbox results later
on.

* Modify snap-refresh-revert jobs

- Reboot commands are moved inside the snapd/snap-[refresh|revert]-*
jobs. This is to make sure the device is rebooted right after the
command is issued (with previous version, some other jobs might be run
between the refresh/revert command and the reboot command, making the
test results unreliable)
- As a result, the snapd/snap-[refresh|revert]-* are flagged as noreturn
- Because Checkbox does not capture the outputs of noreturn jobs, their
outputs are stored in the $PLAINBOX_SESSION_SHARE directory using `tee`
- Reboot jobs are replaced with attachment jobs to upload the outputs
along with the submissions

* Replace reboot jobs with log attachment jobs in snap-refresh-revert test plan

Fix CHECKBOX-1264

* Fix snapd unit tests

* Add unit test to improve coverage

* Fix unit test

* Add template-id to log-attach jobs

Following the work done in commit
b8befd8ffff1f2dd6aee0485cf8a1c24a374f2c0, add template-id to the new
templates that don't have them set yet.