Code formatting using black (infra) (#1163)
* Set line length to 79 when calling black

black uses pyproject.toml to store its configuration. Update the
different projects in the monorepo to match line length of 79 when
calling black.

See black documentation[1] for more information.

[1]
https://black.readthedocs.io/en/stable/usage_and_configuration/the_basics.html#configuration-via-a-file

* Update contributing guide to mention black

* Add black GitHub Action

* Ignore vendorized code when running black

* Format all the codebase using black

Command used:

    black . --line-length 79 --extend-exclude "/vendor/"


* Ignore E203 (Whitespace before ':')

As explained in this issue[1], E203[2] is not PEP 8 compliant so it
should be ignored when running flake8.

Black is now being used in the codebase, and it takes care of the
formatting.

[1] https://github.com/psf/black/issues/315
[2] https://www.flake8rules.com/rules/E203.html

* Ignore flake8 complaints for a few specific lines in the base provider

Although Black is configured at 79 chars line length, it keeps a few
lines to 80, 81 or 82 chars. Flake8 will complain about this, but we can
ignore these.