Add backoff and jitter to connection function
Following feedback from team, reworking the connection function to use a
random delay, using a backoff and a jitter. Each new attempt will wait
longer than the previous one (up to 60 seconds per attempt).

The jitter is here to prevent the test from choking the infrastructure
if many devices are trying to run this test at the same moment.