make miscellanea/check-gpu-driver a cert-blocker
Make miscellanea/check-gpu-driver to be a blocker. In a recent sutton case
(PX-Ref4), the alpha image was using the nvidia dkms module since the signed
module was not ready yet.

miscellanea/check-gpu-driver had figured out the issue, but ever since it is
non-blocker, QA and PM failed to pay much attention to it, and nearly GMed the
platform with it.

Discussed this with Michael and Phoenix, we all agree to make this test case a
cert blocker.

JIRA: OEX86-299
Signed-off-by: Atlas Yu <atlas.yu@canonical.com>