Build executables into `bin/` and `data/`
This commit changes the gpu-setup script behaviour to build the cuda-samples
and gpu-burn projects inside the `build/bin` directory, then copy them
out into the `bin/` and (the necessary data files) into `data/`. For the
cuda-samples executables to work, they need access to the data files,
but they do not take the path to the data dir as an argument; to
circumvent this limitation, I have made wrapper scripts that copy the
necessary file into the temporary working directory that checkbox
creates.

Because of the change in build behaviour, the `gpu-setup` script now runs
mostly as a regular user (to avoid permission issues when cleaning
directories/builds). The expected operation now is to run `./manage.py build`
instead of running the `gpu-setup.sh` script itself. This is more inline
with what is done with the other providers.