Add manifest for interactive LED jobs
Instead of asking the tester to manually skip tests when a device does
not have a given LED (for instance no Microphone Mute LED), this commit
introduces manifest entries for all the related LED jobs.

This way, the tester defines what LEDs the device has before running the
tests, and tests that are not required due to a lack of specific LED are
automatically skipped.