New match feature (new) (#1422)
Re-running an entire test plan is very time consuming, sometimes we have to do it to only fix a handful of jobs and this stresses our lab and makes us waste time. One solution on the Checkbox side is to have a new configuration that allows us to run a few specific tests.

Similarly to how `exclude=` works, a new `match` section is added to the [test selection] configuration that does the opposite of exclude. Namely: Given a bootstrapped test plan run all tests that match the pattern(s) or that the matching tests depend on.

Documentation and Metabox tests are included.