Consider ID_INPUT_POINTINGSTICK as touchpad (BugFix)
Most resistive touchpads don't support ABS coordination nor multitouch.
When both capabilities are absent, the touchpad gets tagged as
ID_INPUT_MOUSE.

Since there's no such device as I2C mouse, if the touchpad is on I2C
bus, it will also get tagged as ID_INPUT_POINTINGSTICK.

To really distinguish a true pointing stick and a resistive touchpad,
the parser needs to find the slibling node created by HID Report
Descriptor and ensure it's a "touchpad" (i.e. ABS or MT capability).
It's not easy to traverse to a sibling node in the current parser, so
simply treat ID_INPUT_POINTINGSTICK as resistive touchpad to resolve the
issue.