Refactor open_connection() and secured_connection()
Both function are doing exactly the same thing, the only difference
being the command used.

Create a connection() function, and call it from open_connection() and
secured_connection().

Compared to the previous implementation, the connection() function is
more linear and does not rely on boolean return codes to decide what to
do. Instead, it's trying to do things and expect exceptions to be raised
in case of problem (this is why it's using subprocess.run(...,
check=True) for instance)