Use the retry decorator in WiFi tests (Bugfix) (#1488)
* Add the function name when running a retry

This helps investigating issues when more than one functions are using
the retry decorator.

* Rename retry mock function

* Refactor open_connection() and secured_connection()

Both function are doing exactly the same thing, the only difference
being the command used.

Create a connection() function, and call it from open_connection() and
secured_connection().

Compared to the previous implementation, the connection() function is
more linear and does not rely on boolean return codes to decide what to
do. Instead, it's trying to do things and expect exceptions to be raised
in case of problem (this is why it's using subprocess.run(...,
check=True) for instance)

* Let exceptions bubble up instead of catching them

* Use the retry decorator

Parts of the wifi_nmcli_test.py script was implementing a basic retry
mechanism. Remove this to use the retry decorator instead.

In addition, raise exceptions rather than trying to catching them, or
using boolean return codes.

* Adjust unit tests

* Add mock_retry to the unit tests for the main function

This is to avoid waiting for the actual retry decorator implementation.

* Fix unit test to work with Python 3.5

* Fix long line that failed the flake8 tests

* Fix perform_ping_test to raise an exception

* Improve unit test coverage

* Update providers/base/bin/wifi_nmcli_test.py

Co-authored-by: Massimiliano <massimiliano.girardi@canonical.com>

* Update providers/base/bin/wifi_nmcli_test.py

Co-authored-by: Massimiliano <massimiliano.girardi@canonical.com>

* Cleanup tests and unused variables to make automated checks happy

* Replace run() with check_call()

* Add contributors in script header

* Fix more unit tests

---------

Co-authored-by: Massimiliano <massimiliano.girardi@canonical.com>