Switch from mobilebroadband to wwan nested parts for cert test plans (bugfix) (#1496)
* Switch from mobilebroadband to wwan nested parts for cert test plans

Following QA changes in #821¹ regarding the replacement of
mobilebroadband-cert-.* nested part for their wwan-.* counterparts, the
same is applied for:

- client-cert-desktop-24-04 (it was missed in #821 due to bad timing)
- client-cert-odm-desktop-2?-04 (to match what's being tested by the QA
team)

client-cert-odm-desktop-18-04 was not included due to the fact that
Ubuntu 18.04 has seen EOL and therefore its related test plan is not
actively maintained anymore.

¹ https://github.com/canonical/checkbox/pull/821

* Add after-suspend-wwan-manual for homogeneity

* Remove superseded jobs from [after-suspend-]wwan-manual nested parts

wwan/detect-manual and wwan/check-sim-present-manual have an automated
version available. wwan/gsm-connection-interrupted-manual depends on the
manual versions of the detection and SIM verification jobs, and is
therefore removed as well, as we have an automated connection job.