Add manifest to check grade status of an image
In our testing lab, we need to provision some devices with Ubuntu Core
images that are set to use the 'dangerous' grade.

This will fail the `image/model-grade` test.

Add a new `dangerous_grade_core_image` manifest to decide whether or not
we should be running this test.

It will be run by default (when there is no manifest entry, its value is
set to 'False'), but by setting it to True, the test will be skipped.