Extract compilation from gpu-setup.
Now gpu-setup is merely installing dependencies for the provider. Ideally, these dependencies will get pulled in by the snap or deb package, and the system will be able to compile the tests.

TOOD: With AMD tests being added, this Makefile will need to be skipped if the system being used does not have NVIDIA GPUs.