Update the list of rejected jobs if match function is used
The jobs that are being pruned by the `match` function should be added
to the list of rejected jobs, otherwise the metadata generated at the
end of a test run is incorrect.

Because the list of rejected jobs is initially empty but can be altered
in several places (`run_alternate_selection()` and
`finish_bootstrap()`), it is updated in these functions instead of being
replaced by another list.

The metadata that states if a test plan has been modified
(`custom_joblist`) should also be set to True in these cases.

Fix #1568