Add "internal microphone" manifest entry
Devices without an internal microphone should not run the
`audio/alsa_record_playback_internal` job, as it requires one.

This issue was detected when reviewing test results to issue a
certificate for a desktop PC. The `audio/alsa_record_playback_internal`
job was marked as skipped, but it is a certification-blocker, so it
triggered some alerts.