Fix netplan config uc16 (BugFix) (#1598)
Wireless tests started failing for Ubuntu Core 16 after a change in Checkbox was introduced to use a different netplan config. The auth field was added to the netplan config, which is incompatible with the Ubuntu Core 16 netplan config format.

The Ubuntu version is now checked, and the old netplan configuration type is used for Ubuntu 16.