Fail snap refresh/revert jobs if the snap_update_test.py script fails
Adding `set -eo pipefail` in the snapd refresh/revert jobs that initiate
a reboot ensures that:

- any failure in the `snap_update_test.py` is not masked by the pipe
command (this is achieved by `set -o pipefail`)
- any failure at any step of the `command:` field just fails the whole
job (this is ensured by `set -e`). This is important to avoid the
`reboot` command from being issued if something goes wrong in the
script.

Fix #1615