Indicate a failed job if an exception is caught when running snapd command
When running a snapd revert or refresh command, a SnapdRequestError
exception might be raised (for example when the refresh command failed
to run within the given time).

If this is the case, the script fails, and if it is executed as part of
a Checkbox run, a `__result` file is generated and stored in the shared
session directory to record the fact that this job failed.

When the session is resumed, Checkbox will automatically resume the
session and mark this test as failed.