Fix GPGPU provider distribution (New) (#1542)
* Add CUDA repos to snap repos

* Add gitignore in gpgpu/data

These files are created by the system being tested. They shouldn't be pushed to github.

* Extract compilation from gpu-setup.

Now gpu-setup is merely installing dependencies for the provider. Ideally, these dependencies will get pulled in by the snap or deb package, and the system will be able to compile the tests.

TOOD: With AMD tests being added, this Makefile will need to be skipped if the system being used does not have NVIDIA GPUs.

* Generalize gpgpu/data gitignore

* Add installation of CUDA as postinst

* Try adding Ubuntu-repo CUDA

I'm curious to see if the packaging will succeed with what we have in the Ubuntu repositories. I'm skeptical, though.

* Fix typo

* Install gpu-burn and cuda-samples in postinst

The projects are downloaded to `/opt/` but the binaries are symlinked to `/usr/local/sbin`.

Need to test if the symlinks are able to find the necessary data files.

* Postinst is a bash script

* GPGPU provider only works with amd64 for now

* Re-use REPO_URL variable

* Exit postinst on failure

* Fix typos in postinst

* Avoid failures when temporary files don't exist

* Try without cleaning up temp files

* Is set -e the problem?

* Split `postinst` into `preinst` and `postinst`

* Add AMD repos to snaps

* Setup repos in postinst.

The preinst and postinst solution did not work. preinst does not run before the dependency handling occurs.

* Add rocm-validation-suite to snap dependencies

* Test using snap GPGPU tools package

* Fix typo in snapcraft.yaml

* Slim down checkbox-gpgpu-tools snap

* Install snaps in postinst

* Fix requires snaps in gpgpu jobs

* Add gpu-burn and rvs snaps

* Update checkbox-gpgpu-tools snapcraft.yaml

* Fix jobs.pxu

* Add necessary plugs to checkbox-gpgpu-tools snap

* Add snapd to depends

This should be installed already, but might as well

* Delete data gitignore

These are no longer going to be living inside the data directory

* There are stable releases of gpu-burn and rvs

* Match spacing in jobs.pxu

* Moving cuda-samples snap code to own repo

* Snaps should have correct permissions now

* Use (expected) cuda-samples snap

* Add cuda-samples snap

(not yet owned by us, but should work soon)

* snapd should already be installed

* Install bin_dir

* Suggest ubuntu-drivers-common

This is suggested because servers with NVIDIA GPUs should be able to use
this utility to install the appropriate driver.

* Remove GPGPU tools from stage-snaps

We will figure these out in a follow-up PR.

* tee output

* Shellcheck fixes