Include image information in Checkbox submission (New) (#1460)
This creates a new system_information collector that collects the URL if possible (namely, if ubuntu-report is installed on the machine and if the DCD string is there and valid). This safely fails in unexpected situations, reporting the error. This also stores (in the stderr of the collector) the full report non-destructively, so in any situation we can go back and re-interpret the data.

This is inspired by PC Sanity provider's `get-image-url.sh` script.