Document additional dependencies (Infra) (#1705)
Document additional development dependencies

The pkg-config and libsystemd-dev packages are needed for building
checkbox, but they are not in the documentation. Without them, the
command "python3 -m pip install -e ." (which appears later in the
documentation) fails. To solve this, add them to the command that
installs system packages for checkbox development.