replace check_intel.sh with .py script & remove some tests
The removed tests (from PE) were not testing the right things,
mainly when Nvidia GPU is enabled before / during Intel GPU is
enabled, picking and counting both towards the test, and not
giving a reliable result.  We do test for explicit success of
the rollout of the daemonsets, which is similar to what we do
for Nvidia GPU enabling with microk8s.