add check_intel.py to enable and verify Intel GPU
The enabling is done using the `enable_intel.sh` also added in this
commit, and is copied almost verbatim from `check_intel.sh`, which in
turn was an almost verbatim implementation of instructions from the DSS
documentation.  The `enable_intel.sh` script was too involved to convert
to a Python script.