use TF device-model queues in GH workflow matrix
by removing the specific device identifier, we open up the job to be
picked by multiple machines of the same model in TF that listen to a
common queue based on their model name.  This helps run more of the jobs
in the matrix in parallel, but has the downside that some of the picked
queues may be flaky.  We had previously set the queue nams to specific
devices, but some of them have also turned out to be flaky over time.