Make sure OS resource job gets the proper version on devices running Ubuntu Core (bugfix) (#1748)
* Make operating system resource job compatible with snap

When installed on a Ubuntu Core system, the /etc/[lsb|os]-release files
accessed by Checkbox might be the ones provided by the core snap
installed. As a result, if the core snap does not match the version of
Ubuntu Core, it will result in wrong information provided. For instance,
if checkbox20 is installed on a device running UC22, the information
returned by the lsb_resource job will be "20.04" when it should be
"22.04".

To fix this, if run from a snap, the resource job script will now look
at the content from /var/lib/snapd/hostfs/etc/os-release to make sure it
retrieves the correct information.

Moreover, the script now uses data from the `os-release` file instead of
the `lsb-release` file, as the former is standardized on all
systemd-enabled systems.

Fix C3-1000

* Create a new `os` resource job and soft deprecate `lsb`

Since the resource job that retrieves information about the operating
system is now using info from the `os-release` file, it is renamed `os`,
and the original `lsb` is kept, but marked as deprecated.

* Make graphics_card_resource.py use functions from os_resource.py

graphics_card_resource.py script is modified to use the functions
available from the os_resource.py script to prevent code duplication and
errors with UC systems.

* fix black formatting

* Fix flake8 tests

* Improve test coverage of graphics_card_resource

* Feedback from PR