Run `ethernet/ping_interface` unless hidden manifest set (e.g. lab runs) (new) (#1709)
This PR does two things:

1. Setup a new hidden manifest to prevent running `ethernet/ping_interface` if the hidden manifest is set
2. Setup GitHub Actions (on the Checkbox side) to check if new hidden manifests are being created, and if so, if devices in the lab have been setup accordingly

Keep reading for more information on both.

New hidden manifest:

* Create a hidden manifest to ignore disconnected Ethernet interfaces

* Condition run of ethernet/ping_interface to the hidden manifest

Only run ethernet/ping_interface if the
_ignore_disconnected_ethernet_interface hidden manifest is set to false.
Since this is the case by default, this test will always be run unless
the hidden manifest is set to True (which we can use in the lab where
only one of the device's Ethernet interface is connected).

* Add ethernet/ping-with-any-cable-interface to the ethernet-automated testplan


GitHub Action to check for missing manifests in devices setup in the lab:

* Add PR validation action to check for missing manifests

Trigger the "check missing manifests" action in the DUT config
repository if new hidden manifests are detected.

* Add comparison scripts

* Add note in contributing guide about hidden manifests