Eth hotplug check cable and routable before pinging (Bugfix) (#1694)
Some of the machines (most of the IOT devices) can't ping after plugging the eth cable.
Some of the machines (most of the PC devices) can ping after plugging the eth cable.

In Desktop image, the networking renderer is Networkmanager. It will delay to turn down the eth interface and keep the ip (will not send req to DHCP to get the IP)  when unpluggin the cable.
- If the tester remove the cable and connect the cable immediatly, the test case can be pass (This is the work around method for the original script)
- If the tester remove the cable, make sure the NetworkMangager trun the eth interface down, and reconnect the cable back. In this situation, some of the device will get the ip slowly and this test case will fail. But it really can get the original ip and available to ping.

In Server/UC image, the networking renderer is systemd-networkd. It will immediatly turn the interface down when unplugging the cable.
- If the tester remove the cable and reconnect the cable immedatly, this test case will still fail. But it really can get the original ip and available to ping.

Changes

Need to make sure the networking renderer actually not routable if remove the cable; it actually routable if reconnect the cable before ping.

Flow
1. Check interface cable state (make sure this eth interface is existed)
2. Ask Tester to remove the eth cable
3. check cable is remove
4. check network renderer is Not routable
5. Ask Tester to reconnect the eth cable
6. check cable is connect
7. check network renderer is routable
8. Ping the gateway 

Fix #1197
Fix #1260