Default to content from /etc/os-release
Replace `get_release_file_path` function with `get_release_file_content`
that tries to access `/var/lib/snapd/hostfs/etc/os-release`. If this
file doesn't exist, it means we are not calling the script from within a
snap, so it tries to access `/etc/os-release` instead.

`get_release_info` now takes the content of the `os-release` file as
parameter and ignores empty lines.