Replace `check_dss.sh` with a simple wrapper and add `dss/purge` (New) (#1743)
* Add a simple wrapper to run dss commands

* Switch to run_dss.sh for running some dss commands

* Switch to kubectl directly for verifying dss ns

* Switch to checking dss status with inline grep

* Remove obsolete check_dss.sh

* Add timeout while creating notebooks

the command may get stuck if, e.g. notebook pod gets in a crashloop

* Add dss/purge test

* Fix run_dss.sh by making steps explicit

The shebang magic with env did not work out :(
but at least now we are explicit, which is better?

* Fix apt installs not being noninteractive

* Fix apt installs not being noninteractive in job-def

* Enable pseudo-terminal in ssh cmds in job-def

* Run dss commands from /tmp instead of HOME

DSS likes to create a `logs` directory inside the working dir,
so we need to run it from outside the snap's working dir.  There
is no mention of it being critical to maintain these files across
reboots, so we may just as well use /tmp instead of HOME.

* Only unset PYTHONHOME and warn with DSS issue

* Add warning with issue about changing working dir