Remove tests relying on Intel GPU counts
These tests are unreliable because the label gpu.intel.com inaccurately
starts counting NVIDIA GPUs too once the respective plugin is enabled.
The same source of inaccuracy also disallowed us from making accurate
checks for the capacity and allocatable slots available in the cluster.

We are removing this test also because the DSS how-to we have been
following does not ask for it.  See:
<https://github.com/canonical/data-science-stack/blob/e495bacef97a1b6bf8bdb63dca29912be317be8c/docs/how-to/enable-gpus/enable-intel-gpu.rst>