Change nvidia gpu val. to depend on dss/initialize
This job needs to run after either one of the two jobs above it enabling
NVIDIA GPU in the k8s cluster succeed (one is for microk8s, the other is
for Canonical k8s).  We can't use those jobs in 'depends' because then
both of them will have to succeed, which is impossible because only one
of either microk8s or Canonical k8s will be available.

The trick we use here is that we now `depends` on `dss/initialize`,
which must succeed for the whole test-plan to be run anyway, and,
we require that we have an NVIDIA GPU.  This is similar to the `depends`
for the two jobs for microk8s and Canonical k8s.  Then we will have to
be careful that this job is added in the test-plan to ONLY run after
those two jobs.  The difference will then be that this job will not be
skipped if either of the two jobs enabling NVIDIA GPU fail.