Simplify enabling Intel GPU for DSS (New) (#1789)
* Move enabling intel gpu to a standalone scripts

One script enables the Intel GPU plugin, the second verifies the rollout
status of the necessary daemonsets.

* Remove redundant tests to check daemonset rollout

We check rollout of these daemonset in the /install job, and the command
`kubectl rollout status` would not have finished successfully if these
daemonsets did not become ready.  We did that during /install to stick
to the pattern of enabling NVIDIA GPU addon.

* Move checking intel gpu label in-line

* Remove tests relying on Intel GPU counts

These tests are unreliable because the label gpu.intel.com inaccurately
starts counting NVIDIA GPUs too once the respective plugin is enabled.
The same source of inaccuracy also disallowed us from making accurate
checks for the capacity and allocatable slots available in the cluster.

We are removing this test also because the DSS how-to we have been
following does not ask for it.  See:
<https://github.com/canonical/data-science-stack/blob/e495bacef97a1b6bf8bdb63dca29912be317be8c/docs/how-to/enable-gpus/enable-intel-gpu.rst>

* Fix name of script to check nvidia gpu rollout

* Add a combined script for checking GPU rollout

* Use unified script to check GPU rollout

* Remove individual scripts for checking GPU rollout

* Switch to checking rollout of parent daemonset

While this branch is not impacted, but when enabling NVIDIA GPU in
Canonical Kubernetes, the `worker` daemonset can have a version in its
name that we cannot predict.  We are making this update here as well to
avoid merge conflicts or forgetting to update this.

* Update check to verify Intel GPU label

Now we test that at least one node in a poetentially multi-node k8s
cluster has the required label for Intel GPU being available.